requirements.txt
# Core Framework
flask==3.0.0
flask-cors==4.0.0
flask-restful==0.3.10
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1

# ML/AI
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2

# Security & Scanning
python-nmap==0.7.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# CVE and Vulnerability Data
cvss==2.6
nvdlib==0.7.4

# Data Processing
pyyaml==6.0.1
python-dotenv==1.0.0

# API & Authentication
pyjwt==2.8.0
bcrypt==4.1.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
